notification and change minor

+++ app/src/main/java/com/rt04/myapplication/core/data/Spending.kt
@@ -5,7 +5,7 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class Spending(
-    val jumlah : Int? = null,
+    val jumlah : Double? = null,
     val nama : String? = null,
     val tanggal : String? = null,
 ): Parcelable
